+++ CHANGELOG.md
@@ -245,3 +245,4 @@
 - Minor improvements to the Entity Genes datapack loader
 - Entity Genes are now datagenned rather than manually typed
 - The Ender Dragon no longer can give the Basic Gene
+- The info pages that show what entities give what genes now uses an Organic Matter item rather than a DNA Helix item